Use #save! when not handling return value
    user = User.new
    user.name = "John"
    user.save

What is wrong with this piece of code? It will **fail silently** when user
cannot be saved. There will be no single trace of this failure in your
logs and you will spend time wondering: “why there are no users in the
database”. If you expect that data is valid and model should be always
saved successfully, then use bang versions – `save!`, `create!` and so on.
Use save only when you handle the return value:

    if user.save
      # ...
    else
      # ...
    end

Via http://blog.sundaycoding.com/blog/2015/01/31/11-easy-to-fix-ruby-slash-ruby-on-rails-mistakes/